added code C80 to valid codes list, per Dr. Bob Anderson's advice
"With regard to the first issue, C80 was expanded by WHO to include the
4th digits .0 and .9 in 2010.  This change was implemented in Iris at
the same time.  However, in the US, we are behind in implementing the
updates, so we are still using C80.  I think for the algorithm, I think
you could simply add C80 to the valid code list."